--- leetcode/p1.h
@@ -1,15 +1,0 @@
-#ifndef P1_H
-#define P1_H
-
-
-void    p1_two_sum();
-void    p2_Add_Two_Numbers();
-void    p3_Longest_Substring_Without_Repeating_Characters();
-void    p4_Median_of_Two_Sorted_Arrays();
-
-void    p9_PalindromeNumber();
-
-void    p95();
-void    p96();
-
-#endif
