+++ leetcode/problem.h
@@ -1,0 +1,19 @@
+#include "0/p1.h"
+#include "0/p2.h"
+#include "0/p3.h"
+#include "0/p4.h"
+
+#include "0/p9.h"
+
+#include "0/p95.h"
+#include "0/p96.h"
+
+
+#include "6/p669.h"
+
+
+#include "8/p895.h"
+
+
+#include "10/p1007.h"
+
