+++ leetcode/src/0/p1.h
@@ -1,0 +1,8 @@
+#ifndef P1_H
+#define P1_H
+
+
+void    p1_two_sum();
+
+
+#endif
