+++ leetcode/src/0/p2.h
@@ -1,0 +1,8 @@
+#ifndef P2_H
+#define P2_h
+
+
+void    p2_Add_Two_Numbers();
+
+
+#endif
