+++ leetcode/src/0/p4.h
@@ -1,0 +1,8 @@
+#ifndef P4_H
+#define P4_H
+
+
+void    p4_Median_of_Two_Sorted_Arrays();
+
+
+#endif
