+++ leetcode/src/0/p9.h
@@ -1,0 +1,8 @@
+#ifndef P9_H
+#define P9_H
+
+s
+void    p9_PalindromeNumber();
+
+
+#endif
