+++ leetcode/src/10/p1007.cpp
@@ -1,4 +1,4 @@
-#include "p10.h"
+#include "p1007.h"
 
 #include <vector>
 #include <iostream>
