+++ leetcode/src/10/p1007.h
@@ -1,5 +1,5 @@
-#ifndef P10_H
-#define P10_H
+#ifndef P1007_H
+#define P1007_H
 
 
 void p1007_Minimum_Domino_Rotations_For_Equal_Row();
