+++ leetcode/src/6/p669.cpp
@@ -1,4 +1,4 @@
-#include "p6.h"
+#include "p669.h"
 
 #include <iostream>
