+++ leetcode/src/6/p669.h
@@ -1,5 +1,5 @@
-#ifndef P6_H
-#define P6_H
+#ifndef P669_H
+#define P669_H
