+++ leetcode/src/8/p895.h
@@ -1,5 +1,6 @@
-#ifndef P8_H
-#define P8_H
+#ifndef P895_H
+#define P895_H
+
 
 void p895_Maximum_Frequency_Stack();
